dave/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-cors-headers==3.14.0
django-health-check==3.17.0
djangorestframework==3.14.0
dnspython==2.3.0
pymongo==4.3.3
pytz==2023.3
sqlparse==0.4.4
